feat: update markdown template for custom fields option and update tests
Since config is not passed to mardown time entries report, the condition
to show custom fields row is based on the existance of any of them.

If there are no custom fields, or there are custom fields with empty
values, the row is not displayed. However the tests are updated to
match the padding for the `Custom fields` row in the first column.

Signed-off-by: Caleb Trepowski <calebtrepowski@gmail.com>